requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
pyserial==3.5
eventlet==0.33.3
